symlink support and checks in deletes etc

+++ lib/CmsFilesystem.php
@@ -270,22 +270,31 @@ class CmsFilesystem {
 
   function fileRename($fileid,$filename){
     $fileinfo= $this->fileInfo;
-    $fileid = mysql_escape_string($fileid);
-    $filename = mysql_escape_string($filename);
-    $filename = str_replace("\\","",$filename);
-    $filename = str_replace("/","",$filename);
-    $fileinfo = $this->getFileInfo($fileid);
-    $query = "UPDATE wildfire_file set filename=\"$filename\" where id=$fileid";
-    $result = $this->query($query);
-    rename($fileinfo['path'].'/'.$fileinfo['filename'],$fileinfo['path'].'/'.$filename);
+    $file = new WildfireFile($fileid);
+    if($file->primval){
+      $path = $file->path.$file->filename;
+      if(!is_link($path)){
+        $fileid = mysql_escape_string($fileid);
+        $filename = mysql_escape_string($filename);
+        $filename = str_replace("\\","",$filename);
+        $filename = str_replace("/","",$filename);
+        $fileinfo = $this->getFileInfo($fileid);
+        $query = "UPDATE wildfire_file set filename=\"$filename\" where id=$fileid";
+        $result = $this->query($query);
+        rename($fileinfo['path'].'/'.$fileinfo['filename'],$fileinfo['path'].'/'.$filename);
+      }
+    }
+    
   }
 
   function fileDelete($fileid){
     $fileid = mysql_escape_string($fileid);
 		$model = new WildfireFile($fileid);
 		$fileinfo = $this->getFileInfo($fileid);
-		unlink($fileinfo['path'].'/'.$fileinfo['filename']) || $this->error('file error');
-		if($model->id) $mod = $model->delete();
+		if(!is_link($fileinfo['path'].'/'.$fileinfo['filename'])){
+		  unlink($fileinfo['path'].'/'.$fileinfo['filename']) || $this->error('file error');
+		  if($model->id) $mod = $model->delete();
+	  }		
 	  echo "File successfully deleted";
     exit;
   }
@@ -293,21 +302,25 @@ class CmsFilesystem {
   function fileMove($fileid,$path){
   	$fileinfo = $this->fileinfo;
   	$defaultFileStore = $this->defaultFileStore;
-      
-  	$fileid = mysql_escape_string($fileid);
+    $file = new WildfireFile($fileid);
+    if($file->primval){
+      $path = $file->path.$file->filename;
+      if(!is_link($path)){  
+  	    $fileid = mysql_escape_string($fileid);	
+  	    $path = str_replace("//","/",$path);
+      	$path = str_replace("..","",$path);
 	
-  	$path = str_replace("//","/",$path);
-  	$path = str_replace("..","",$path);
-	
-  	$path = mysql_escape_string($path);
-    $fileinfo = $this->getFileInfo($fileid);
-  	$newPath = $this->defaultFileStore.$path;
-  	if(is_dir($newPath)){
-      $query = "UPDATE wildfire_file set path=\"$newPath\",rpath=\"$path\" where id=$fileid";
-  		$result = $this->query($query);
-  		rename($fileinfo['path'].'/'.$fileinfo['filename'],$newPath.'/'.$fileinfo['filename']);
-  		echo "done";
-  	} else $this->error('new directory doesnt exist');
+      	$path = mysql_escape_string($path);
+        $fileinfo = $this->getFileInfo($fileid);
+      	$newPath = $this->defaultFileStore.$path;
+      	if(is_dir($newPath)){
+          $query = "UPDATE wildfire_file set path=\"$newPath\",rpath=\"$path\" where id=$fileid";
+      		$result = $this->query($query);
+      		rename($fileinfo['path'].'/'.$fileinfo['filename'],$newPath.'/'.$fileinfo['filename']);
+      		echo "done";
+      	} else $this->error('new directory doesnt exist');
+    	}
+  	}
   }
 
   function folderRename($path,$name,$newname){
@@ -319,7 +332,7 @@ class CmsFilesystem {
     $currentPath = $this->defaultFileStore.$path.'/'.$name;
     $newPath = $this->defaultFileStore.$path.'/'.$newname;
 
-    if(is_dir($currentPath) && !is_dir($newPath)){
+    if(is_dir($currentPath) && !is_dir($newPath) && !is_link($currentPath)){
 
   	  if(rename($currentPath,$newPath)){
   	    $query = "UPDATE wildfire_file set path=\"$newPath\",rpath=\"$path/$newname\" where path=\"$currentPath\"";
@@ -344,7 +357,7 @@ class CmsFilesystem {
     $userPath = $this->defaultFileStore.$path.'/'.$name;
   	$userNewPath = $this->defaultFileStore.$newpath.'/'.$name;
       
-  	if(is_dir($userPath) && !is_dir($userNewPath)){
+  	if(is_dir($userPath) && !is_dir($userNewPath) && !is_link($currentPath)){
 		
   		if(rename($userPath,$userNewPath)){
   		  $query = "UPDATE wildfire_file set path=\"$userNewPath\",rpath=\"$newpath/$name\" where path=\"$userPath\"";
@@ -358,18 +371,15 @@ class CmsFilesystem {
 
 
   function folderDelete($folder){
-
-  	$folder = mysql_escape_string($folder);
-
-
-  		$deleteDir = $this->defaultFileStore.$folder;
-	
+    $folder = mysql_escape_string($folder);
+    $deleteDir = $this->defaultFileStore.$folder;	
+    if(!is_link($deleteDir)){  	  
   		if($this->deleteDir($deleteDir)){
   			$query = "DELETE from wildfire_file where path like \"$deleteDir\%\"";
   			$result = $this->query($query);
   			echo "ok";
-  		} else echo "oops somethings wrong";
-	
+  		}else echo "oops somethings wrong";
+	  }
   }
 
   function newFolder($name,$path){
